+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - FAQ in the [README](https://github.com/ttous/basic-web-app/blob/master/README.md).
 ### Changed
+- Update [clean-webpack-plugin](https://github.com/johnagan/clean-webpack-plugin) to version 2.0.0.
 - Replace [file-loader](https://github.com/webpack-contrib/file-loader) by [url-loader](https://github.com/webpack-contrib/url-loader).
 
 ## [0.0.6] - 2019-03-01